pyproject.toml
[project]
name = "build-loop"
version = "0.1.0"
description = "Generalized multi-agent build system"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.39.0",
    "pydantic>=2.0",
    "rich>=13.0",
    "python-dotenv>=1.0",
]

[project.scripts]
build-loop = "build_loop.main:main"
